@@ -63,12 +63,12 @@ class MovieTeaserBottomSheet extends StatelessWidget {
           children: [
             Text(
               movieTitle,
-              style: Theme.of(context).textTheme.headline6,
+              style: Theme.of(context).textTheme.titleLarge,
             ),
             const Gap(4),
             Text(
               movieYear,
-              style: Theme.of(context).textTheme.bodyText2,
+              style: Theme.of(context).textTheme.bodyMedium,
             ),
             const Divider(),
             Row(
@@ -164,12 +164,12 @@ class TextOptionButton extends StatelessWidget {
   Widget build(BuildContext context) {
     return TextButton(
       onPressed: onPressed,
-      child: Text(text),
       style: ButtonStyle(
         padding: MaterialStateProperty.all(
           EdgeInsets.zero,
         ),
       ),
+      child: Text(text),
     );
   }
 }
