@@ -82,7 +82,7 @@ class UserPanel extends StatelessWidget {
           Expanded(
             child: Text(
               displayName,
-              style: Theme.of(context).textTheme.headline6,
+              style: Theme.of(context).textTheme.titleLarge,
             ),
           ),
         ],
