@@ -8,7 +8,7 @@ import Foundation
 import cloud_firestore
 import firebase_auth
 import firebase_core
-import path_provider_macos
+import path_provider_foundation
 
 func RegisterGeneratedPlugins(registry: FlutterPluginRegistry) {
   FLTFirebaseFirestorePlugin.register(with: registry.registrar(forPlugin: "FLTFirebaseFirestorePlugin"))
