@@ -1,3 +1,5 @@
+// ignore_for_file: depend_on_referenced_packages
+
 import 'package:film_freund/managers/cache/cache_manager.dart';
 import 'package:film_freund/managers/movies/movie_manager.dart';
 import 'package:film_freund/managers/user/user_manager.dart';
@@ -10,7 +12,6 @@ import 'package:film_freund/services/platform/i_platform_service.dart';
 import 'package:film_freund/services/user/i_user_database.dart';
 import 'package:film_freund/services/uuid/uuid_service.dart';
 import 'package:firebase_core_platform_interface/firebase_core_platform_interface.dart';
-import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:hive/hive.dart';
@@ -57,39 +58,10 @@ typedef Callback = void Function(MethodCall call);
 
 class MethodChannelMocks {
   // taken from https://github.com/FirebaseExtended/flutterfire/blob/master/packages/firebase_auth/firebase_auth/test/mock.dart
-  static void setupFirebase([Callback? customHandlers]) {
+  static void setupFirebase() {
     TestWidgetsFlutterBinding.ensureInitialized();
 
-    MethodChannelFirebase.channel.setMockMethodCallHandler((call) async {
-      if (call.method == 'Firebase#initializeCore') {
-        return [
-          {
-            'name': defaultFirebaseAppName,
-            'options': {
-              'apiKey': '123',
-              'appId': '123',
-              'messagingSenderId': '123',
-              'projectId': '123',
-            },
-            'pluginConstants': {},
-          }
-        ];
-      }
-
-      if (call.method == 'Firebase#initializeApp') {
-        return {
-          'name': call.arguments['appName'],
-          'options': call.arguments['options'],
-          'pluginConstants': {},
-        };
-      }
-
-      if (customHandlers != null) {
-        customHandlers(call);
-      }
-
-      return null;
-    });
+    setupFirebaseCoreMocks();
   }
 
   static void setupPathProvider(PathProviderPlatform instance) {
